backend/pyproject.toml
[project]
name = "project-proojectloop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pandas",
    "groq",
    "python-dotenv",
]

